+++ WeightApp Watch App/Assets.xcassets/AccentColor.colorset/Contents.json
