+++ WeightApp Watch App/Assets.xcassets/AppIcon.appiconset/Contents.json
