+++ WeightApp Watch App/Assets.xcassets/Contents.json
