+++ WeightApp Watch App/Models/CalorieDataModel.swift
