+++ WeightApp Watch App/Models/WeightDataModel.swift
