+++ WeightApp Watch App/Preview Content/Preview Assets.xcassets/Contents.json
