+++ WeightApp Watch App/Views/SettingsView.swift
