+++ WeightApp/Assets.xcassets/AccentColor.colorset/Contents.json
