+++ WeightApp/Assets.xcassets/AppIcon.appiconset/Contents.json
