+++ WeightApp/Assets.xcassets/Contents.json
