+++ WeightApp/Assets.xcassets/Images/Contents.json
