+++ WeightApp/Assets.xcassets/Images/bluetooth.imageset/172308_bluetooth_icon.png
