+++ WeightApp/Assets.xcassets/Images/bluetooth.imageset/Contents.json
