+++ WeightApp/Assets.xcassets/Images/firebase.imageset/Contents.json
