+++ WeightApp/Assets.xcassets/Images/firebase.imageset/pngwing.com-2.png
