+++ WeightApp/Assets.xcassets/Images/generic.imageset/Contents.json
