+++ WeightApp/Assets.xcassets/Images/kfscis.imageset/Contents.json
