+++ WeightApp/Assets.xcassets/Images/logo.imageset/Contents.json
