+++ WeightApp/Assets.xcassets/Images/logo.imageset/Screen Shot 2022-11-07 at 9.16.22 AM.png
