+++ WeightApp/Assets.xcassets/Images/user.imageset/Contents.json
