+++ WeightApp/Modifier/ReusableViews/CloseView.swift
