+++ WeightApp/Preview Content/Preview Assets.xcassets/Contents.json
