+++ WeightApp/SessionServices/SessionUserDetails.swift
