+++ WeightApp/Views/Database/RegisterView.swift
