+++ WeightApp/Views/Main/ScaleView.swift
