+++ WeightApp/Views/Main/Settings/BluetoothPeripherals.swift
@@ -1,0 +1,56 @@
+//
+//  BluetoothPeripherals.swift
+//  WeightApp
+//
+//  Created by DJC on 4/18/23.
+//
+
+import SwiftUI
+import CoreBluetooth
+
+class BluetoothViewModel: NSObject, ObservableObject {
+    private var centralManager: CBCentralManager?
+    private var peripherals: [CBPeripheral] = []
+    @Published var peripheralNames: [String] = []
+    
+    override init() {
+        super.init()
+        self.centralManager = CBCentralManager(delegate: self, queue: .main)
+    }
+}
+
+extension BluetoothViewModel: CBCentralManagerDelegate {
+    func centralManagerDidUpdateState(_ central: CBCentralManager) {
+        if central.state == .poweredOn {
+            self.centralManager?.scanForPeripherals(withServices: nil)
+        }
+    }
+    
+    func centralManager(_ central: CBCentralManager, didDiscover peripheral: CBPeripheral, advertisementData: [String : Any], rssi RSSI: NSNumber) {
+        if !peripherals.contains(peripheral) {
+            self.peripherals.append(peripheral)
+            self.peripheralNames.append(peripheral.name ?? peripheral.description)
+        }
+    }
+}
+
+struct BluetoothPeripherals: View {
+    @ObservedObject private var bluetoothViewModel = BluetoothViewModel()
+    @Environment(\.dismiss) var dismiss
+    
+    var body: some View {
+        VStack {
+            List(bluetoothViewModel.peripheralNames, id: \.self) { peripheral in
+                Text(peripheral)
+            }
+            Spacer()
+            Text("Select a peripheral device")
+                .font(.headline)
+            Button("Press to dismiss") {
+                dismiss()
+            }
+            .font(.title)
+            .padding()
+        }
+    }
+}
